+++ tests/testthat/test-period.R
@@ -39,6 +39,37 @@ test_that("period names to date, first date parses, rest bad",
 
 })
 
+test_that("Weekly periods: mixed separators",
+{
+    x <- c("10-15-2012-11-16-2012", "10/15/2012-11/16/2012",
+           "10/15-2012 11-16/2012")
+    out <- AsDate(x)
+    expect_equal(out[1], out[2])
+    expect_equal(out[1], out[3])
+    expect_equal(format(out, "%m"), rep("10", 3L))
+})
+
+test_that("Weekly periods: character months",
+{
+    x <- c("aUG/11/1869-dec/10/1871", "MARCH/26/1970 jun/30/1970")
+    out <- AsDate(x)
+    expect_equal(format(out, "%m"), c("08", "03"))
+    expect_equal(format(out, "%Y"), c("1869", "1970"))
+})
+
+test_that("Weekly periods: two-digit years",
+{
+    x <- c("may/15/03-sep/20/04", "november/29/12-july/13/13")
+    out <- AsDate(x)
+    expect_equal(format(out, "%d"), c("15", "29"))
+    expect_equal(format(out, "%y"), c("03", "12"))
+
+    x <- c("01/01/03-17/05/04", "29/11/12-10/10/13")
+    out <- AsDate(x)
+    expect_equal(format(out, "%d"), c("01", "29"))
+    expect_equal(format(out, "%y"), c("03", "12"))
+})
+
 test_that("Q-Quarterly dates: diff. start and end years; DS-1652",
 {
     x <- c("Jun-Nov 16", "Dec-May 17", "Jun-Nov 17")
@@ -46,6 +77,39 @@ test_that("Q-Quarterly dates: diff. start and end years; DS-1652",
     expect_equal(format(out,  "%m"), c("06", "12", "06"))
     ## Dec-May 17 needs to get converted to 2016-Dec-01
     expect_equal(format(out,  "%y"), c("16", "16", "17"))
+})
+
+test_that("quarterly periods: numeric months",
+{
+    x <- c("10-12 99", "12-03 00")
+    out <- AsDate(x)
+    expect_equal(format(out, "%m"), c("10", "12"))
+    expect_equal(format(out, "%y"), c("99", "99"))
+})
+
+test_that("quarterly periods: full name months",
+{
+    x <- c("JANUARY-feb 99", "december-JUN 00")
+    out <- AsDate(x)
+    expect_equal(format(out, "%m"), c("01", "12"))
+    expect_equal(format(out, "%y"), c("99", "99"))
+})
+
+test_that("quarterly periods: '/' is supported",
+{
+    x <- c("JANUARY/feb 99", "01/03 00")
+    out <- AsDate(x)
+    expect_equal(format(out, "%m"), c("01", "01"))
+    expect_equal(format(out, "%y"), c("99", "00"))
+})
+
+test_that("quarterly periods: four-digit years supported",
+{
+    x <- c("mar/october 1977", "apr/janUary 1989")
+    out <- AsDate(x)
+    expect_equal(format(out, "%m"), c("03", "04"))
+    expect_equal(format(out, "%y"), c("77", "88"))
+
 })
 
 test_that("Periods",